@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState} from 'react';
 import {
   Modal,
   ScrollView,
@@ -14,17 +14,20 @@ import Button from './Button';
 // UTILS
 import {height} from '../utils/constants';
 
-class TextModal extends React.Component {
-  constructor(props) {
-    super(props);
+const TextModal = ({
+  field,
+  visible,
+  toggleModal,
+  type,
+  check,
+  customStyles,
+  styleName,
+  defaultTextModal,
+}) => {
+  const [disabledButton, setDisabledButton] = useState(true);
+  const [screenHeight, setScreenHeight] = useState(0);
 
-    this.state = {
-      disabledButton: true,
-      screenHeight: 0,
-    };
-  }
-
-  isCloseToBottom = ({layoutMeasurement, contentOffset, contentSize}) => {
+  const isCloseToBottom = ({layoutMeasurement, contentOffset, contentSize}) => {
     const paddingToBottom = 20;
     return (
       layoutMeasurement.height + contentOffset.y >=
@@ -32,88 +35,76 @@ class TextModal extends React.Component {
     );
   };
 
-  onContentSizeChange = (contentWidth, contentHeight) => {
-    this.setState({screenHeight: contentHeight});
+  const onContentSizeChange = (contentWidth, contentHeight) => {
+    setScreenHeight(contentHeight);
   };
 
-  render() {
-    const {disabledButton, screenHeight} = this.state;
-    const {
-      field,
-      visible,
-      toggleModal,
-      type,
-      check,
-      customStyles,
-      styleName,
-      defaultTextModal,
-    } = this.props;
-    const scrollEnabled = screenHeight > (height - 100);
-    const disBtn = disabledButton && scrollEnabled;
-    return (
-      <Modal
-        visible={visible}
-        animationType="fade"
-        transparent
-        onRequestClose={() => toggleModal()}>
+  const scrollEnabled = screenHeight > (height - 100);
+  const disBtn = disabledButton && scrollEnabled;
+
+  return (
+    <Modal
+      visible={visible}
+      animationType="fade"
+      transparent
+      onRequestClose={() => toggleModal()}>
+      <View
+        style={StyleSheet.flatten([
+          styles.textModalContainer,
+          customStyles.textModalContainer,
+          customStyles[`textModalContainer${styleName}`],
+        ])}>
         <View
           style={StyleSheet.flatten([
-            styles.textModalContainer,
-            customStyles.textModalContainer,
-            customStyles[`textModalContainer${styleName}`],
+            styles.textModalWrapper,
+            customStyles.textModalWrapper,
+            customStyles[`textModalWrapper${styleName}`],
           ])}>
           <View
             style={StyleSheet.flatten([
-              styles.textModalWrapper,
-              customStyles.textModalWrapper,
-              customStyles[`textModalWrapper${styleName}`],
+              styles.closeModal,
+              customStyles.closeModal,
+              customStyles[`closeModal${styleName}`],
             ])}>
-            <View
-              style={StyleSheet.flatten([
-                styles.closeModal,
-                customStyles.closeModal,
-                customStyles[`closeModal${styleName}`],
-              ])}>
-              <TouchableOpacity onPress={() => toggleModal()}>
-                <Icon name="close" size={20} />
-              </TouchableOpacity>
-            </View>
-            <ScrollView
-              scrollEnabled={scrollEnabled}
-              contentContainerStyle={styles.textModalScrollView}
-              onScroll={({nativeEvent}) => {
-                if (this.isCloseToBottom(nativeEvent)) {
-                  this.setState({disabledButton: false});
-                }
-              }}
-              scrollEventThrottle={400}
-              onContentSizeChange={this.onContentSizeChange}>
-              <Text>{field.text || defaultTextModal}</Text>
-            </ScrollView>
-            <View>
-              <Button
-                disabled={disBtn}
-                onPress={() => check(true)}
-                buttonLabel={field.buttonAcceptLabel}
-                styleName={styleName}
-                customStyles={{
-                  ...customStyles,
-                  buttonContainer: {
-                    ...customStyles.buttonContainer,
-                    ...styles[
-                      disBtn ? 'acceptButtonDisabled' : 'acceptButtonActive'
-                    ],
-                  },
-                }}>
-                {'form.accept'}
-              </Button>
-            </View>
+            <TouchableOpacity onPress={() => toggleModal()}>
+              <Icon name="close" size={20} />
+            </TouchableOpacity>
+          </View>
+          <ScrollView
+            scrollEnabled={scrollEnabled}
+            contentContainerStyle={styles.textModalScrollView}
+            onScroll={({nativeEvent}) => {
+              if (isCloseToBottom(nativeEvent)) {
+                setDisabledButton(false);
+              }
+            }}
+            scrollEventThrottle={400}
+            onContentSizeChange={onContentSizeChange}>
+            <Text>{field.text || defaultTextModal}</Text>
+          </ScrollView>
+          <View>
+            <Button
+              disabled={disBtn}
+              onPress={() => check(true)}
+              buttonLabel={field.buttonAcceptLabel}
+              styleName={styleName}
+              customStyles={{
+                ...customStyles,
+                buttonContainer: {
+                  ...customStyles.buttonContainer,
+                  ...styles[
+                    disBtn ? 'acceptButtonDisabled' : 'acceptButtonActive'
+                  ],
+                },
+              }}>
+              {'form.accept'}
+            </Button>
           </View>
         </View>
-      </Modal>
-    );
-  }
-}
+      </View>
+    </Modal>
+  );
+};
 
 const styles = StyleSheet.create({
   textModalContainer: {
