@@ -2,36 +2,33 @@ import React from 'react';
 import {StyleSheet, TouchableOpacity, View, Text} from 'react-native';
 import PropTypes from 'prop-types';
 
-class Button extends React.PureComponent {
-  render() {
-    const {onPress, customStyles, styleName, buttonLabel} = this.props;
-    return (
-      <TouchableOpacity
-        onPress={onPress}
+const Button = ({onPress, customStyles, styleName, buttonLabel}) => {
+  return (
+    <TouchableOpacity
+      onPress={onPress}
+      style={StyleSheet.flatten([
+        styles.buttonContainer,
+        customStyles.buttonContainer,
+        customStyles[`buttonContainer${styleName}`],
+      ])}>
+      <View
         style={StyleSheet.flatten([
-          styles.buttonContainer,
-          customStyles.buttonContainer,
-          customStyles[`buttonContainer${styleName}`],
+          styles.buttonContent,
+          customStyles.buttonContent,
+          customStyles[`buttonContent${styleName}`],
         ])}>
-        <View
+        <Text
           style={StyleSheet.flatten([
-            styles.buttonContent,
-            customStyles.buttonContent,
-            customStyles[`buttonContent${styleName}`],
+            styles.buttonText,
+            customStyles.buttonText,
+            customStyles[`buttonText${styleName}`],
           ])}>
-          <Text
-            style={StyleSheet.flatten([
-              styles.buttonText,
-              customStyles.buttonText,
-              customStyles[`buttonText${styleName}`],
-            ])}>
-            {buttonLabel}
-          </Text>
-        </View>
-      </TouchableOpacity>
-    );
-  }
-}
+          {buttonLabel}
+        </Text>
+      </View>
+    </TouchableOpacity>
+  );
+};
 
 const styles = StyleSheet.create({
   buttonContainer: {
