@@ -3,25 +3,22 @@ import {StyleSheet, TouchableOpacity} from 'react-native';
 import PropTypes from 'prop-types';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 
-class Checkbox extends React.PureComponent {
-  render() {
-    const {customStyles, onPress, value, field, customIcon} = this.props;
-    const {icon, iconChecked, iconSize, iconColor} = field;
-    const iconName = icon ? icon : 'checkbox-blank-outline';
-    const iconNameChecked = iconChecked ? iconChecked : 'check-box-outline';
-    const FormIcon = customIcon || Icon;
-    return (
-      <TouchableOpacity
-        style={StyleSheet.flatten([styles.checkbox, customStyles.checkbox])}
-        onPress={() => onPress()}>
-        <FormIcon
-          name={value ? iconNameChecked : iconName}
-          size={iconSize || 30}
-          color={iconColor}
-        />
-      </TouchableOpacity>
-    );
-  }
+const Checkbox = ({customStyles, onPress, value, field, customIcon}) => {
+  const {icon, iconChecked, iconSize, iconColor} = field;
+  const iconName = icon ? icon : 'checkbox-blank-outline';
+  const iconNameChecked = iconChecked ? iconChecked : 'check-box-outline';
+  const FormIcon = customIcon || Icon;
+  return (
+    <TouchableOpacity
+      style={StyleSheet.flatten([styles.checkbox, customStyles.checkbox])}
+      onPress={() => onPress()}>
+      <FormIcon
+        name={value ? iconNameChecked : iconName}
+        size={iconSize || 30}
+        color={iconColor}
+      />
+    </TouchableOpacity>
+  );
 }
 
 const styles = StyleSheet.create({
