@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {
   StyleSheet,
   View,
@@ -11,112 +11,102 @@ import Checkbox from './Checkbox';
 import TextModal from './TextModal';
 import ErrorMessage from './ErrorMessage';
 
-class CheckboxField extends React.Component {
-  state = {
-    modalVisible: false,
-    value: false,
-  };
+const CheckboxField = ({
+  field,
+  customStyles,
+  label,
+  styleName,
+  errors,
+  defaultTextModal,
+  customIcon,
+  defaultValue,
+  onChange,
+}) => {
+  const {name, openModal = false} = field;
 
-  componentDidMount() {
-    const {defaultValue} = this.props;
+  const [modalVisible, setModalVisible] = useState(false);
+  const [value, setValue] = useState(false);
+
+  useEffect(() => {
     if (defaultValue) {
-      this.setState({
-        value: defaultValue,
-      });
+      setValue(defaultValue);
     }
-  }
+  }, []);
 
-  selectAction = () => {
-    const {value} = this.state;
-    const {field} = this.props;
+  const selectAction = () => {
     const {openModal = false} = field;
     if (value) {
-      this.check(false);
+      check(false);
     } else if (openModal) {
-      this.setState(prevState => ({modalVisible: !prevState.modalVisible}));
+      setModalVisible(!modalVisible);
     } else {
-      this.check(true);
+      check(true);
     }
   };
 
-  check = val => {
-    const {onChange} = this.props;
+  const check = val => {
     onChange(val);
-    this.setState({value: val, modalVisible: false});
+    setValue(val);
+    setModalVisible(false);
   };
 
-  render() {
-    const {modalVisible, value} = this.state;
-    const {
-      field,
-      customStyles,
-      label,
-      styleName,
-      reference,
-      errors,
-      defaultTextModal,
-      customIcon,
-    } = this.props;
-    const {name, openModal = false} = field;
-    return (
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.checkboxContainer,
+        customStyles.checkboxContainer,
+        customStyles[`checkboxContainer${styleName}`],
+      ])}>
       <View
         style={StyleSheet.flatten([
-          styles.checkboxContainer,
-          customStyles.checkboxContainer,
-          customStyles[`checkboxContainer${styleName}`],
+          styles.checkboxWrapper,
+          customStyles.checkboxWrapper,
+          customStyles[`checkboxWrapper${styleName}`],
         ])}>
-        <View
+        <Checkbox
+          field={field}
+          value={value}
+          customIcon={customIcon}
+          onPress={() => selectAction()}
+        />
+        <TouchableOpacity
+          onPress={() => selectAction()}
           style={StyleSheet.flatten([
-            styles.checkboxWrapper,
-            customStyles.checkboxWrapper,
-            customStyles[`checkboxWrapper${styleName}`],
+            styles.checkboxTextWrapper,
+            customStyles.checkboxTextWrapper,
+            customStyles[`checkboxTextWrapper${styleName}`],
           ])}>
-          <Checkbox
-            field={field}
-            ref={reference}
-            value={value}
-            customIcon={customIcon}
-            onPress={() => this.selectAction()}
-          />
-          <TouchableOpacity
-            onPress={() => this.selectAction()}
+          <Text
             style={StyleSheet.flatten([
-              styles.checkboxTextWrapper,
-              customStyles.checkboxTextWrapper,
-              customStyles[`checkboxTextWrapper${styleName}`],
+              styles.checkboxText,
+              customStyles.checkboxText,
+              customStyles[`checkboxText${styleName}`],
             ])}>
-            <Text
-              style={StyleSheet.flatten([
-                styles.checkboxText,
-                customStyles.checkboxText,
-                customStyles[`checkboxText${styleName}`],
-              ])}>
-              {label}
-            </Text>
-          </TouchableOpacity>
-          {openModal && (
-            <TextModal
-              visible={modalVisible}
-              check={this.check}
-              toggleModal={this.selectAction}
-              field={field}
-              defaultTextModal={defaultTextModal}
-              styleName={styleName}
-              customStyles={customStyles}
-            />
-          )}
-        </View>
-        {errors[name] && (
-          <ErrorMessage
-            error={errors[name]}
-            name={styleName}
+            {label}
+          </Text>
+        </TouchableOpacity>
+        {openModal && (
+          <TextModal
+            visible={modalVisible}
+            check={check}
+            toggleModal={selectAction}
+            field={field}
+            defaultTextModal={defaultTextModal}
+            styleName={styleName}
             customStyles={customStyles}
           />
         )}
       </View>
-    );
-  }
-}
+      {errors[name] && (
+        <ErrorMessage
+          error={errors[name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   checkboxContainer: {
@@ -135,18 +125,22 @@ const styles = StyleSheet.create({
 
 CheckboxField.defaultProps = {
   onPress: () => {},
+  onChange: () => {},
   customStyles: {},
   defaultValue: null,
   label: '',
+  defaultTextModal: '',
 };
 
 CheckboxField.propTypes = {
   field: PropTypes.object.isRequired,
   styleName: PropTypes.string.isRequired,
-  reference: PropTypes.func.isRequired,
   onPress: PropTypes.func,
   customStyles: PropTypes.object,
   label: PropTypes.string,
+  defaultTextModal: PropTypes.string,
+  defaultValue: PropTypes.any,
+  onChange: PropTypes.func,
 };
 
 export default CheckboxField;
