@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {Text, View, StyleSheet} from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import PropTypes from 'prop-types';
@@ -6,122 +6,111 @@ import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 // COMPONENTS
 import ErrorMessage from './ErrorMessage';
 
-class DatePicker extends React.Component {
-  state = {
-    date: new Date(),
-    show: false,
-  };
+const DatePicker = ({
+  field,
+  label,
+  onChange,
+  errors,
+  customStyles,
+  currentLocale,
+  styleName,
+  customIcon,
+  defaultValue,
+}) => {
+  const [date, setDate] = useState(new Date());
+  const [show, setShow] = useState(false);
 
-  componentDidMount() {
-    const {defaultValue} = this.props;
+  useEffect(() => {
     if (defaultValue) {
-      this.setState({date: new Date(...defaultValue.split('-'))});
+      setDate(new Date(...defaultValue.split('-')));
     }
-  }
+  }, []);
 
-  render() {
-    const {date, show} = this.state;
-    const {
-      field,
-      label,
-      reference,
-      onChange,
-      errors,
-      customStyles,
-      currentLocale,
-      styleName,
-      customIcon,
-    } = this.props;
-    const FormIcon = customIcon || Icon;
-    return (
+  const FormIcon = customIcon || Icon;
+
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.datePickerContainer,
+        customStyles.datePickerContainer,
+        customStyles[`datePickerContainer${styleName}`],
+      ])}>
+      {label && (
+        <Text
+          style={StyleSheet.flatten([
+            styles.datePickerLabel,
+            customStyles.datePickerLabel,
+            customStyles[`datePickerLabel${styleName}`],
+          ])}>
+          {label}
+        </Text>
+      )}
       <View
         style={StyleSheet.flatten([
-          styles.datePickerContainer,
-          customStyles.datePickerContainer,
-          customStyles[`datePickerContainer${styleName}`],
+          styles.datePickerTextWrapper,
+          customStyles.datePickerTextWrapper,
+          customStyles[`datePickerTextWrapper${styleName}`],
+          (field.icon || field.iconLeft || field.iconRight) && {
+            paddingHorizontal: 5,
+          },
         ])}>
-        {label && (
-          <Text
-            style={StyleSheet.flatten([
-              styles.datePickerLabel,
-              customStyles.datePickerLabel,
-              customStyles[`datePickerLabel${styleName}`],
-            ])}>
-            {label}
-          </Text>
-        )}
-        <View
-          style={StyleSheet.flatten([
-            styles.datePickerTextWrapper,
-            customStyles.datePickerTextWrapper,
-            customStyles[`datePickerTextWrapper${styleName}`],
-            (field.icon || field.iconLeft || field.iconRight) && {
-              paddingHorizontal: 5,
-            },
-          ])}>
-          {(field.icon || field.iconLeft) && (
-            <FormIcon
-              name={field.icon || field.iconLeft}
-              size={field.iconSize || 20}
-              color={field.iconColor || '#000'}
-            />
-          )}
-          <Text
-            style={StyleSheet.flatten([
-              styles.datePickerText,
-              customStyles.datePickerText,
-              customStyles[`datePickerText${styleName}`],
-              (field.icon || field.iconLeft || field.iconRight) && {
-                marginHorizontal: 5,
-              },
-            ])}
-            onPress={() =>
-              this.setState(prevState => ({show: !prevState.show}))
-            }>
-            {date.toLocaleDateString(currentLocale)}
-          </Text>
-          {field.iconRight && (
-            <FormIcon
-              name={field.iconRight}
-              size={field.iconSize || 20}
-              color={field.iconColor || '#000'}
-            />
-          )}
-        </View>
-        {show && (
-          <DateTimePicker
-            ref={reference}
-            value={date}
-            minimumDate={
-              field.minimumDate
-                ? new Date(...field.minimumDate.split('-'))
-                : null
-            }
-            maximumDate={
-              field.maximumDate
-                ? new Date(...field.maximumDate.split('-'))
-                : new Date()
-            }
-            onChange={(event, date) => {
-              this.setState(prevState => ({
-                show: !prevState.show,
-                date: date || prevState.date,
-              }));
-              onChange(date);
-            }}
+        {(field.icon || field.iconLeft) && (
+          <FormIcon
+            name={field.icon || field.iconLeft}
+            size={field.iconSize || 20}
+            color={field.iconColor || '#000'}
           />
         )}
-        {errors[field.name] && (
-          <ErrorMessage
-            error={errors[field.name]}
-            name={styleName}
-            customStyles={customStyles}
+        <Text
+          style={StyleSheet.flatten([
+            styles.datePickerText,
+            customStyles.datePickerText,
+            customStyles[`datePickerText${styleName}`],
+            (field.icon || field.iconLeft || field.iconRight) && {
+              marginHorizontal: 5,
+            },
+          ])}
+          onPress={() => setShow(!show)}>
+          {date.toLocaleDateString(currentLocale)}
+        </Text>
+        {field.iconRight && (
+          <FormIcon
+            name={field.iconRight}
+            size={field.iconSize || 20}
+            color={field.iconColor || '#000'}
           />
         )}
       </View>
-    );
-  }
-}
+      {show && (
+        <DateTimePicker
+          value={date}
+          minimumDate={
+            field.minimumDate
+              ? new Date(...field.minimumDate.split('-'))
+              : null
+          }
+          maximumDate={
+            field.maximumDate
+              ? new Date(...field.maximumDate.split('-'))
+              : new Date()
+          }
+          onChange={(event, newDate) => {
+            setShow(!show);
+            setDate(newDate || date);
+            onChange(newDate);
+          }}
+        />
+      )}
+      {errors[field.name] && (
+        <ErrorMessage
+          error={errors[field.name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   datePickerContainer: {
@@ -169,7 +158,6 @@ DatePicker.propTypes = {
   field: PropTypes.object.isRequired,
   styleName: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  reference: PropTypes.func.isRequired,
   label: PropTypes.string,
   value: PropTypes.instanceOf(Date),
   errors: PropTypes.object,
