@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {Text, View, StyleSheet, Modal, TouchableOpacity} from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import PropTypes from 'prop-types';
@@ -8,169 +8,158 @@ import ErrorMessage from './ErrorMessage';
 // UTILS
 import {height, width} from '../utils/constants';
 
-class DatePicker extends React.Component {
-  state = {
-    date: new Date(),
-    newDate: new Date(),
-    show: false,
-  };
+const DatePicker = ({
+  field,
+  label,
+  onChange,
+  errors,
+  customStyles,
+  currentLocale,
+  styleName,
+  customIcon,
+  defaultValue,
+}) => {
+  const [date, setDate] = useState(new Date());
+  const [newDate, setNewDate] = useState(new Date());
+  const [show, setShow] = useState(false);
 
-  componentDidMount() {
-    const {defaultValue} = this.props;
+  useEffect(() => {
     if (defaultValue) {
-      this.setState({
-        newDate: new Date(...defaultValue.split('-')),
-        date: new Date(...defaultValue.split('-')),
-      });
+      setDate(new Date(...defaultValue.split('-')));
+      setNewDate(new Date(...defaultValue.split('-')));
     }
-  }
+  }, []);
 
-  toggleModal = () => {
-    this.setState(prevState => ({
-      show: !prevState.show,
-      newDate: prevState.date,
-    }));
+  const toggleModal = () => {
+    setShow(!show);
+    setNewDate(date);
   };
 
-  render() {
-    const {date, newDate, show} = this.state;
-    const {
-      field,
-      label,
-      reference,
-      onChange,
-      errors,
-      customStyles,
-      currentLocale,
-      styleName,
-      customIcon,
-    } = this.props;
-    const FormIcon = customIcon || Icon;
-    return (
-      <View
-        ref={reference}
-        style={StyleSheet.flatten([
-          styles.datePickerContainer,
-          customStyles.datePickerContainer,
-          customStyles[`datePickerContainer${styleName}`],
-        ])}>
-        {label && (
-          <Text
-            style={StyleSheet.flatten([
-              styles.datePickerLabel,
-              customStyles.datePickerLabel,
-              customStyles[`datePickerLabel${styleName}`],
-            ])}>
-            {label}
-          </Text>
-        )}
-        <View
+  const FormIcon = customIcon || Icon;
+
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.datePickerContainer,
+        customStyles.datePickerContainer,
+        customStyles[`datePickerContainer${styleName}`],
+      ])}>
+      {label && (
+        <Text
           style={StyleSheet.flatten([
-            styles.datePickerTextWrapper,
-            customStyles.datePickerTextWrapper,
-            customStyles[`datePickerTextWrapper${styleName}`],
-            (field.icon || field.iconLeft || field.iconRight) && {
-              paddingHorizontal: 5,
-            },
+            styles.datePickerLabel,
+            customStyles.datePickerLabel,
+            customStyles[`datePickerLabel${styleName}`],
           ])}>
-          {(field.icon || field.iconLeft) && (
-            <FormIcon
-              name={field.icon || field.iconLeft}
-              size={field.iconSize || 20}
-              color={field.iconColor || '#000'}
-            />
-          )}
-          <Text
-            style={StyleSheet.flatten([
-              styles.datePickerText,
-              customStyles.datePickerText,
-              customStyles[`datePickerText${styleName}`],
-              (field.icon || field.iconLeft || field.iconRight) && {
-                marginHorizontal: 5,
-              },
-            ])}
-            onPress={() => this.toggleModal()}>
-            {date.toLocaleDateString(currentLocale)}
-          </Text>
-          {field.iconRight && (
-            <FormIcon
-              name={field.iconRight}
-              size={field.iconSize || 20}
-              color={field.iconColor || '#000'}
-            />
-          )}
-        </View>
-        <Modal
-          animationType="fade"
-          transparent
-          visible={show}
-          onRequestClose={() => this.toggleModal()}>
-          <TouchableOpacity
-            style={StyleSheet.flatten([
-              styles.datePickerModalOverlay,
-              customStyles.datePickerModalOverlay,
-              customStyles[`datePickerModalOverlay${styleName}`],
-            ])}
-            activeOpacity={1}
-            onPress={() => this.toggleModal()}
+          {label}
+        </Text>
+      )}
+      <View
+        style={StyleSheet.flatten([
+          styles.datePickerTextWrapper,
+          customStyles.datePickerTextWrapper,
+          customStyles[`datePickerTextWrapper${styleName}`],
+          (field.icon || field.iconLeft || field.iconRight) && {
+            paddingHorizontal: 5,
+          },
+        ])}>
+        {(field.icon || field.iconLeft) && (
+          <FormIcon
+            name={field.icon || field.iconLeft}
+            size={field.iconSize || 20}
+            color={field.iconColor || '#000'}
           />
-          <View
-            style={StyleSheet.flatten([
-              styles.datePickerModalContainer,
-              customStyles.datePickerModalContainer,
-              customStyles[`datePickerModalContainer${styleName}`],
-            ])}>
-            <View
-              style={StyleSheet.flatten([
-                styles.datePickerModalWrapper,
-                customStyles.datePickerModalWrapper,
-                customStyles[`datePickerModalWrapper${styleName}`],
-              ])}>
-              <View
-                style={StyleSheet.flatten([
-                  styles.datePickerModalConfirmContainer,
-                  customStyles.datePickerModalConfirmContainer,
-                  customStyles[`datePickerModalConfirmContainer${styleName}`],
-                ])}>
-                <TouchableOpacity
-                  onPress={() => {
-                    this.setState({date: newDate});
-                    onChange(newDate);
-                    this.toggleModal();
-                  }}>
-                  <Text>Done</Text>
-                </TouchableOpacity>
-              </View>
-              <DateTimePicker
-                value={newDate}
-                minimumDate={
-                  field.minimumDate
-                    ? new Date(...field.minimumDate.split('-'))
-                    : null
-                }
-                maximumDate={
-                  field.maximumDate
-                    ? new Date(...field.maximumDate.split('-'))
-                    : new Date()
-                }
-                onChange={(event, d) => {
-                  this.setState({newDate: d});
-                }}
-              />
-            </View>
-          </View>
-        </Modal>
-        {errors[field.name] && (
-          <ErrorMessage
-            error={errors[field.name]}
-            name={styleName}
-            customStyles={customStyles}
+        )}
+        <Text
+          style={StyleSheet.flatten([
+            styles.datePickerText,
+            customStyles.datePickerText,
+            customStyles[`datePickerText${styleName}`],
+            (field.icon || field.iconLeft || field.iconRight) && {
+              marginHorizontal: 5,
+            },
+          ])}
+          onPress={() => toggleModal()}>
+          {date.toLocaleDateString(currentLocale)}
+        </Text>
+        {field.iconRight && (
+          <FormIcon
+            name={field.iconRight}
+            size={field.iconSize || 20}
+            color={field.iconColor || '#000'}
           />
         )}
       </View>
-    );
-  }
-}
+      <Modal
+        animationType="fade"
+        transparent
+        visible={show}
+        onRequestClose={() => toggleModal()}>
+        <TouchableOpacity
+          style={StyleSheet.flatten([
+            styles.datePickerModalOverlay,
+            customStyles.datePickerModalOverlay,
+            customStyles[`datePickerModalOverlay${styleName}`],
+          ])}
+          activeOpacity={1}
+          onPress={() => toggleModal()}
+        />
+        <View
+          style={StyleSheet.flatten([
+            styles.datePickerModalContainer,
+            customStyles.datePickerModalContainer,
+            customStyles[`datePickerModalContainer${styleName}`],
+          ])}>
+          <View
+            style={StyleSheet.flatten([
+              styles.datePickerModalWrapper,
+              customStyles.datePickerModalWrapper,
+              customStyles[`datePickerModalWrapper${styleName}`],
+            ])}>
+            <View
+              style={StyleSheet.flatten([
+                styles.datePickerModalConfirmContainer,
+                customStyles.datePickerModalConfirmContainer,
+                customStyles[`datePickerModalConfirmContainer${styleName}`],
+              ])}>
+              <TouchableOpacity
+                onPress={() => {
+                  setDate(newDate);
+                  onChange(newDate);
+                  toggleModal();
+                }}>
+                <Text>Done</Text>
+              </TouchableOpacity>
+            </View>
+            <DateTimePicker
+              value={newDate}
+              minimumDate={
+                field.minimumDate
+                  ? new Date(...field.minimumDate.split('-'))
+                  : null
+              }
+              maximumDate={
+                field.maximumDate
+                  ? new Date(...field.maximumDate.split('-'))
+                  : new Date()
+              }
+              onChange={(event, d) => {
+                setNewDate(d);
+              }}
+            />
+          </View>
+        </View>
+      </Modal>
+      {errors[field.name] && (
+        <ErrorMessage
+          error={errors[field.name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   datePickerContainer: {
@@ -228,7 +217,6 @@ DatePicker.propTypes = {
   field: PropTypes.object.isRequired,
   styleName: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  reference: PropTypes.func.isRequired,
   label: PropTypes.string,
   value: PropTypes.instanceOf(Date),
   errors: PropTypes.object,
