@@ -2,27 +2,24 @@ import React from 'react';
 import {Text, StyleSheet, View} from 'react-native';
 import PropTypes from 'prop-types';
 
-class ErrorMessage extends React.PureComponent {
-  render() {
-    const {error, name, customStyles} = this.props;
-    return (
-      <View
+const ErrorMessage = ({error, name, customStyles}) => {
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.errorMessageContainer,
+        customStyles.errorMessageContainer,
+        customStyles[`errorMessageContainer${name}`],
+      ])}>
+      <Text
         style={StyleSheet.flatten([
-          styles.errorMessageContainer,
-          customStyles.errorMessageContainer,
-          customStyles[`errorMessageContainer${name}`],
+          styles.errorMessageText,
+          customStyles.errorMessageText,
+          customStyles[`errorMessageText${name}`],
         ])}>
-        <Text
-          style={StyleSheet.flatten([
-            styles.errorMessageText,
-            customStyles.errorMessageText,
-            customStyles[`errorMessageText${name}`],
-          ])}>
-          {error.message}
-        </Text>
-      </View>
-    );
-  }
+        {error.message}
+      </Text>
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
