@@ -19,12 +19,12 @@ import FormGroupLabel from './FormGroupLabel';
 import RadioButtons from './RadioButtons';
 import TextOnly from './TextOnly';
 
-export default ({
+const Form = ({
   formConfig,
   mode = 'onSubmit',
   onSubmit,
   onChangeCustom = (field, value) => value,
-  customStyles,
+  customStyles = {},
   currentLocale,
   customIcon,
   defaultValues = {},
@@ -32,26 +32,27 @@ export default ({
   defaultTextModals = {},
   showPasswordIconName = 'eye-outline',
   hidePasswordIconName = 'eye-off',
+  forceLocale = null
 }) => {
-  let deviceLanguage = null;
-  let lang = null;
-  if (Platform.OS === 'ios') {
-    lang = NativeModules.SettingsManager.settings.AppleLanguages[0];
-    deviceLanguage = lang.split('-')[0];
-  } else {
-    lang = NativeModules.I18nManager.localeIdentifier;
-    deviceLanguage = lang.split('_')[0];
+  let deviceLanguage = forceLocale;
+  if (!forceLocale) {
+    let lang = null;
+    if (Platform.OS === 'ios') {
+      lang = NativeModules.SettingsManager.settings.AppleLanguages[0];
+      deviceLanguage = lang.split('-')[0];
+    } else {
+      lang = NativeModules.I18nManager.localeIdentifier;
+      deviceLanguage = lang.split('_')[0];
+    }
   }
   const validations = () => {
     const validator = {};
     if (formConfig.groups instanceof Array !== true) {
       throw new Error('Configuration expects groups key to be a list of objects');
-      return;
     }
     formConfig.groups.forEach(g => {
       if (g.children instanceof Array !== true) {
         throw new Error('Configuration expects groups key to consist of list of children objects');
-        return;
       }
       g.children.forEach(f => {
         if (f.validations) {
@@ -110,17 +111,24 @@ export default ({
   // const watchFields = watch(fields.reduce((acc, val) => acc.concat(val), []));
   const watchFields = watch();
 
-  // NOTE: Use for add default values
   useEffect(() => {
+    formConfig.groups.forEach(g => {
+      if (g.children instanceof Array !== true) {
+        throw new Error('Configuration expects groups key to consist of list of children objects');
+      }
+      g.children.forEach(f => {
+        register({name: f.name});
+      });
+    });
+    // NOTE: Use for add default values
     Object.keys(defaultValues).forEach(key => {
       setValue(key, defaultValues[key]);
     });
-  }, [defaultValues, register, setValue]);
+  }, [register]);
 
   const formComponents = () => {
     if (formConfig.groups instanceof Array !== true) {
       throw new Error('Configuration expects groups key to be a list of objects');
-      return;
     }
     return formConfig.groups.map((g, index) => (
       <View key={index}>
@@ -133,7 +141,6 @@ export default ({
             }
           />
         )}
-
         {g.children instanceof Array === true && g.children.map(f => {
           const styleName = f.name
             .split(/_|-/i)
@@ -175,7 +182,6 @@ export default ({
                 customIcon={customIcon}
                 showPasswordIconName={showPasswordIconName}
                 hidePasswordIconName={hidePasswordIconName}
-                reference={() => register({name: f.name})}
                 defaultValue={defaultValues[f.name]}
                 changeBackgroundOnFocus={f.changeBackgroundOnFocus}
                 onChange={text => {
@@ -201,7 +207,6 @@ export default ({
                 label={label}
                 defaultValue={defaultValues[f.name]}
                 currentLocale={currentLocale || deviceLanguage}
-                reference={() => register({name: f.name})}
                 onChange={date => {
                   const customDate = onChangeCustom(f.name, date);
                   setValue(f.name, customDate, f.require);
@@ -226,7 +231,6 @@ export default ({
                 label={label}
                 defaultValue={defaultValues[f.name]}
                 customIcon={customIcon}
-                reference={() => register({name: f.name})}
                 items={f.items || defaultSelectValues[f.name] || []}
                 onChange={date => {
                   const customDate = onChangeCustom(f.name, date);
@@ -249,7 +253,6 @@ export default ({
                 label={label}
                 defaultValue={defaultValues[f.name]}
                 defaultTextModal={defaultTextModals[f.name]}
-                reference={() => register({name: f.name})}
                 onChange={text => {
                   const customValue = onChangeCustom(f.name, text);
                   setValue(f.name, customValue, f.require);
@@ -268,7 +271,6 @@ export default ({
                 label={label}
                 defaultValue={defaultValues[f.name]}
                 placeholder={placeholder}
-                reference={() => register({name: f.name})}
                 onChange={text => {
                   const customText = onChangeCustom(f.name, text);
                   setValue(f.name, customText, f.require);
@@ -301,7 +303,7 @@ export default ({
           return null;
         })}
       </View>
-    ))
+    ));
   };
   const showSubmit =
     formConfig.showSubmit === undefined ? true : formConfig.showSubmit;
@@ -358,3 +360,5 @@ const styles = StyleSheet.create({
     fontWeight: '800',
   },
 });
+
+export default Form;
