@@ -2,25 +2,22 @@ import React from 'react';
 import {Text, StyleSheet, View} from 'react-native';
 import PropTypes from 'prop-types';
 
-class FormGroupLabel extends React.PureComponent {
-  render() {
-    const {label, customStyles} = this.props;
-    return (
-      <View
+const FormGroupLabel = ({label, customStyles}) => {
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.formGroupLabelContainer,
+        customStyles.formGroupLabelContainer,
+      ])}>
+      <Text
         style={StyleSheet.flatten([
-          styles.formGroupLabelContainer,
-          customStyles.formGroupLabelContainer,
+          styles.formGroupLabelText,
+          customStyles.formGroupLabelText,
         ])}>
-        <Text
-          style={StyleSheet.flatten([
-            styles.formGroupLabelText,
-            customStyles.formGroupLabelText,
-          ])}>
-          {label}
-        </Text>
-      </View>
-    );
-  }
+        {label}
+      </Text>
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
