@@ -1,5 +1,6 @@
-import React from 'react';
+import React, {useState} from 'react';
 import {
+  Keyboard,
   StyleSheet,
   Text,
   TextInput,
@@ -11,43 +12,46 @@ import PropTypes from 'prop-types';
 import ErrorMessage from './ErrorMessage';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 
-class InputText extends React.Component {
-  state = {
-    showPassword: false,
-    changeBg: false,
-  };
+const InputText = ({
+  showPasswordIconColor,
+  showPasswordIconSize,
+  showPasswordIconName,
+  hidePasswordIconName,
+  onBlur,
+  field,
+  label,
+  onChange,
+  errors,
+  iconSize,
+  iconColor,
+  customStyles,
+  styleName,
+  placeholder,
+  changeBackgroundOnFocus,
+  defaultValue,
+  customIcon,
+}) => {
+  const [showPassword, setShowPassword] = useState(false);
+  const [changeBg, setChangeBg] = useState(false);
 
-  onBlur = async () => {
-    const {onBlur} = this.props;
-    const {changeBg} = this.state;
+  const onBlurInput = async () => {
     if (changeBg) {
-      this.setState({changeBg: false});
+      setChangeBg(false);
     }
 
     await onBlur();
   };
 
-  renderShowPassword = () => {
-    const {showPassword} = this.state;
-    const {
-      showPasswordIconColor,
-      showPasswordIconSize,
-      customIcon,
-      showPasswordIconName,
-      hidePasswordIconName,
-    } = this.props;
-    const FormIcon = customIcon || Icon;
-    showPasswordIcon = showPasswordIconName;
-    hidePasswordIcon = hidePasswordIconName;
+  const FormIcon = customIcon || Icon;
+
+  const renderShowPassword = () => {
     return (
       <TouchableOpacity
         onPress={() => {
-          this.setState(prevState => ({
-            showPassword: !prevState.showPassword,
-          }));
+          setShowPassword(!showPassword);
         }}>
         <FormIcon
-          name={showPassword ? showPasswordIcon : hidePasswordIcon}
+          name={showPassword ? showPasswordIconName : hidePasswordIconName}
           size={showPasswordIconSize}
           color={showPasswordIconColor}
         />
@@ -55,116 +59,104 @@ class InputText extends React.Component {
     );
   };
 
-  changeBackground = changeBackgroundOnFocus => {
+  const changeBackground = changeBackgroundOnFocus => {
     if (changeBackgroundOnFocus) {
-      this.setState({changeBg: true});
+      setChangeBg(true);
     }
   };
 
-  render() {
-    const {showPassword, changeBg} = this.state;
-    const {
-      field,
-      label,
-      onChange,
-      reference,
-      errors,
-      iconSize,
-      iconColor,
-      customStyles,
-      styleName,
-      placeholder,
-      changeBackgroundOnFocus,
-      defaultValue,
-      customIcon,
-      showPasswordIconName,
-      hidePasswordIconName,
-    } = this.props;
-    const FormIcon = customIcon || Icon;
-    let iconName = null;
-    if (field.icon) {
-      iconName = field.icon.name !== undefined ? field.icon.name : field.icon;
-    }
-    if (field.iconLeft) {
-      iconName =
-        field.iconLeft.name !== undefined
-          ? field.iconLeft.name
-          : field.iconLeft;
-    }
-    if (field.iconRight) {
-      iconName =
-        field.iconRight.name !== undefined
-          ? field.iconRight.name
-          : field.iconRight;
-    }
-    return (
+  let iconName = null;
+  if (field.icon) {
+    iconName = field.icon.name !== undefined ? field.icon.name : field.icon;
+  }
+  if (field.iconLeft) {
+    iconName =
+      field.iconLeft.name !== undefined
+        ? field.iconLeft.name
+        : field.iconLeft;
+  }
+  if (field.iconRight) {
+    iconName =
+      field.iconRight.name !== undefined
+        ? field.iconRight.name
+        : field.iconRight;
+  }
+
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.inputTextContainer,
+        customStyles.inputTextContainer,
+        customStyles[`inputTextContainer${styleName}`],
+      ])}>
+      {label && (
+        <Text
+          style={StyleSheet.flatten([
+            styles.inputTextLabel,
+            customStyles.inputTextLabel,
+            customStyles[`inputTextLabel${styleName}`],
+          ])}>
+          {label}
+        </Text>
+      )}
       <View
         style={StyleSheet.flatten([
-          styles.inputTextContainer,
-          customStyles.inputTextContainer,
-          customStyles[`inputTextContainer${styleName}`],
+          styles.inputTextWrapper,
+          customStyles.inputTextWrapper,
+          customStyles[`inputTextWrapper${styleName}`],
+          (field.icon || field.iconLeft || field.iconRight) && {
+            paddingHorizontal: 5,
+          },
         ])}>
-        {label && (
-          <Text
-            style={StyleSheet.flatten([
-              styles.inputTextLabel,
-              customStyles.inputTextLabel,
-              customStyles[`inputTextLabel${styleName}`],
-            ])}>
-            {label}
-          </Text>
+        {(field.icon || field.iconLeft) && (
+          <FormIcon name={iconName} size={iconSize} color={iconColor} />
         )}
-        <View
+        <TextInput
+          {...field}
+          secureTextEntry={field.secureTextEntry && !showPassword}
+          placeholder={placeholder}
+          value={defaultValue}
+          onChangeText={text => {
+            onChange(text);
+          }}
+          onSubmitEditing={()=> Keyboard.dismiss()}
+          blurOnSubmit={false}
+          onBlur={() => onBlurInput()}
+          onFocus={() => changeBackground(changeBackgroundOnFocus)}
           style={StyleSheet.flatten([
-            styles.inputTextWrapper,
-            customStyles.inputTextWrapper,
-            customStyles[`inputTextWrapper${styleName}`],
+            styles.inputText,
+            customStyles.inputText,
+            customStyles[`inputText${styleName}`],
             (field.icon || field.iconLeft || field.iconRight) && {
-              paddingHorizontal: 5,
+              marginHorizontal: 5,
             },
-          ])}>
-          {(field.icon || field.iconLeft) && (
-            <FormIcon name={iconName} size={iconSize} color={iconColor} />
-          )}
-          <TextInput
-            {...field}
-            secureTextEntry={field.secureTextEntry && !showPassword}
-            ref={reference}
-            placeholder={placeholder}
-            value={defaultValue}
-            onChangeText={text => onChange(text)}
-            onBlur={() => this.onBlur()}
-            onFocus={() => this.changeBackground(changeBackgroundOnFocus)}
-            style={StyleSheet.flatten([
-              styles.inputText,
-              customStyles.inputText,
-              customStyles[`inputText${styleName}`],
-              (field.icon || field.iconLeft || field.iconRight) && {
-                marginHorizontal: 5,
-              },
-              changeBg && [
-                styles.inputTextOnFocusBackground,
-                customStyles.inputTextOnFocusBackground,
-                customStyles[`inputTextOnFocusBackground${styleName}`],
-              ],
-            ])}
-          />
-          {field.type === 'password' && this.renderShowPassword()}
-          {field.type !== 'password' && field.iconRight && (
-            <FormIcon name={iconName} size={iconSize} color={iconColor} />
-          )}
-        </View>
-        {errors[field.name] && (
-          <ErrorMessage
-            error={errors[field.name]}
-            name={styleName}
-            customStyles={customStyles}
-          />
+            changeBg && [
+              styles.inputTextOnFocusBackground,
+              customStyles.inputTextOnFocusBackground,
+              customStyles[`inputTextOnFocusBackground${styleName}`],
+            ],
+            field.editable != null && field.editable === false && [
+              styles.inputTextNotEditable,
+              customStyles.inputTextNotEditable,
+              customStyles[`inputTextNotEditable${styleName}`],
+            ]
+          ])}
+        />
+        {field.type === 'password' && renderShowPassword()}
+        {field.type !== 'password' && field.iconRight && (
+          <FormIcon name={iconName} size={iconSize} color={iconColor} />
         )}
       </View>
-    );
-  }
-}
+      {errors[field.name] && (
+        <ErrorMessage
+          error={errors[field.name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   inputTextContainer: {
@@ -189,6 +181,9 @@ const styles = StyleSheet.create({
     backgroundColor: 'gray',
     color: 'white',
   },
+  inputTextNotEditable: {
+    backgroundColor: 'gray',
+  }
 });
 
 InputText.defaultProps = {
@@ -211,7 +206,6 @@ InputText.propTypes = {
   field: PropTypes.object.isRequired,
   styleName: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  reference: PropTypes.func.isRequired,
   label: PropTypes.string,
   onBlur: PropTypes.func,
   errors: PropTypes.object,
@@ -223,7 +217,6 @@ InputText.propTypes = {
   hidePasswordIconName: PropTypes.string,
   placeholder: PropTypes.string,
   customStyles: PropTypes.object,
-
 };
 
 export default InputText;
