@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {Text, StyleSheet, View, TouchableOpacity} from 'react-native';
 import PropTypes from 'prop-types';
 import HTMLView from 'react-native-htmlview';
@@ -6,44 +6,39 @@ import NestedScrollView from 'react-native-nested-scroll-view';
 // COMPONENTS
 import ErrorMessage from './ErrorMessage';
 
-class RadioButtons extends React.PureComponent {
-  constructor() {
-    super();
+const RadioButtons = ({
+  items,
+  label,
+  styleName,
+  customStyles,
+  errors,
+  field,
+  description,
+  defaultValue,
+  currentLocale,
+  onChange,
+}) => {
+  const [value, setValue] = useState();
 
-    this.state = {
-      value: null,
-    };
-  }
-
-  componentDidMount() {
-    const {defaultValue} = this.props;
+  useEffect(() => {
     if (defaultValue) {
-      this.setState({value: defaultValue});
+      setValue(defaultValue);
     }
-  }
+  }, []);
 
-  renderRadioButton = (item, index) => {
-    const {
-      currentLocale,
-      onChange,
-      customStyles,
-      styleName,
-      reference,
-    } = this.props;
-    const {value} = this.state;
+  const renderRadioButton = (item, index) => {
     const label =
       item.label instanceof Object ? item.label[currentLocale] : item.label;
     return (
       <TouchableOpacity
         key={index}
-        ref={reference}
         style={StyleSheet.flatten([
           styles.radioButtonWrapper,
           customStyles.radioButtonWrapper,
           customStyles[`radioButtonWrapper${styleName}`],
         ])}
         onPress={() => {
-          this.setState({value: item.value});
+          setValue(item.value);
           onChange(item.value);
         }}>
         <View
@@ -74,61 +69,50 @@ class RadioButtons extends React.PureComponent {
     );
   };
 
-  render() {
-    const {
-      items,
-      label,
-      styleName,
-      customStyles,
-      errors,
-      field,
-      description,
-    } = this.props;
-    return (
-      <View
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.radioButtonsContainer,
+        customStyles.radioButtonsContainer,
+        customStyles[`radioButtonsContainer${styleName}`],
+      ])}>
+      <Text
         style={StyleSheet.flatten([
-          styles.radioButtonsContainer,
-          customStyles.radioButtonsContainer,
-          customStyles[`radioButtonsContainer${styleName}`],
+          styles.radioButtonLabel,
+          customStyles.radioButtonLabel,
+          customStyles[`radioButtonLabel${styleName}`],
         ])}>
-        <Text
+        {label}
+      </Text>
+      {description && (
+        <NestedScrollView
           style={StyleSheet.flatten([
-            styles.radioButtonLabel,
-            customStyles.radioButtonLabel,
-            customStyles[`radioButtonLabel${styleName}`],
+            styles.radioButtonTextArea,
+            customStyles.radioButtonTextArea,
+            customStyles[`radioButtonTextArea${styleName}`],
           ])}>
-          {label}
-        </Text>
-        {description && (
-          <NestedScrollView
+          <HTMLView value={description} />
+          {/* <Text
             style={StyleSheet.flatten([
-              styles.radioButtonTextArea,
-              customStyles.radioButtonTextArea,
-              customStyles[`radioButtonTextArea${styleName}`],
+              styles.radioButtonDescription,
+              customStyles.radioButtonDescription,
+              customStyles[`radioButtonDescription${styleName}`],
             ])}>
-            <HTMLView value={description} />
-            {/* <Text
-              style={StyleSheet.flatten([
-                styles.radioButtonDescription,
-                customStyles.radioButtonDescription,
-                customStyles[`radioButtonDescription${styleName}`],
-              ])}>
-              {description}
-            </Text> */}
-          </NestedScrollView>
-        )}
-        {items.map((item, index) => this.renderRadioButton(item, index))}
-        {errors[field.name] && (
-          <ErrorMessage
-            error={errors[field.name]}
-            name={styleName}
-            customStyles={customStyles}
-          />
-        )}
-      </View>
-    );
-  }
-}
+            {description}
+          </Text> */}
+        </NestedScrollView>
+      )}
+      {items.map((item, index) => renderRadioButton(item, index))}
+      {errors[field.name] && (
+        <ErrorMessage
+          error={errors[field.name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   radioButtonWrapper: {
