@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {
   Picker,
   Text,
@@ -14,206 +14,187 @@ import ErrorMessage from './ErrorMessage';
 // UTILS
 import {height, width} from '../utils/constants';
 
-class SelectPicker extends React.Component {
-  state = {
-    value: null,
-    valueIndex: null,
-    selectedValue: null,
-    selectedIndex: null,
-    show: false,
-  };
+const SelectPicker = ({
+  field,
+  label,
+  onChange,
+  errors,
+  items,
+  customStyles,
+  styleName,
+  currentLocale,
+  confirmLabel,
+  customIcon,
+  placeholder,
+  defaultValue
+}) => {
+  const [value, setValue] = useState(null);
+  const [valueIndex, setValueIndex] = useState(null);
+  const [selectedValue, setSelectedValue] = useState(null);
+  const [selectedIndex, setSelectedIndex] = useState(null);
+  const [show, setShow] = useState(false);
 
-  componentDidMount() {
-    const {defaultValue, items, onChange} = this.props;
+  useEffect(() => {
     if (defaultValue) {
       const index = items.map(i => i.value).indexOf(defaultValue);
       onChange(defaultValue);
-      this.setState({
-        value: defaultValue,
-        valueIndex: index,
-        selectedValue: defaultValue,
-        selectedIndex: index,
-      });
+      setValue(defaultValue);
+      setValueIndex(index);
+      setSelectedValue(defaultValue);
+      setSelectedIndex(index);
     }
-  }
+  }, []);
 
-  toggleModal = (type = 'out') => {
-    const {onChange, items} = this.props;
-    const {selectedValue, show} = this.state;
+  const toggleModal = (type = 'out') => {
     if (show && !selectedValue && type === 'done') {
       onChange(items[0].value);
-      this.setState({
-        value: items[0].value,
-        valueIndex: 0,
-      });
+      setValue(items[0].value);
+      setValueIndex(0);
     }
-    this.setState(prevState => ({
-      show: !prevState.show,
-      selectedValue: prevState.value,
-      selectedIndex: prevState.valueIndex,
-    }));
+    setShow(!show);
+    setSelectedValue(value);
+    setSelectedIndex(valueIndex);
   };
 
-  render() {
-    const {selectedIndex, selectedValue, show, valueIndex} = this.state;
-    const {
-      field,
-      label,
-      reference,
-      onChange,
-      errors,
-      items,
-      customStyles,
-      styleName,
-      currentLocale,
-      confirmLabel,
-      customIcon,
-      placeholder,
-    } = this.props;
-    const FormIcon = customIcon || Icon;
-    return (
+  const FormIcon = customIcon || Icon;
+
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.selectPickerContainer,
+        customStyles.selectPickerContainer,
+        customStyles[`selectPickerContainer${styleName}`],
+      ])}>
+      {label && (
+        <Text
+          style={StyleSheet.flatten([
+            styles.selectPickerLabel,
+            customStyles.selectPickerLabel,
+            customStyles[`selectPickerLabel${styleName}`],
+          ])}>
+          {label}
+        </Text>
+      )}
       <View
-        ref={reference}
         style={StyleSheet.flatten([
-          styles.selectPickerContainer,
-          customStyles.selectPickerContainer,
-          customStyles[`selectPickerContainer${styleName}`],
+          styles.selectPickerTextWrapper,
+          customStyles.selectPickerTextWrapper,
+          customStyles[`selectPickerTextWrapper${styleName}`],
         ])}>
-        {label && (
+        <TouchableOpacity
+          style={StyleSheet.flatten([
+            styles.selectPickerTextTouch,
+            customStyles.selectPickerTextTouch,
+            customStyles[`selectPickerTextTouch${styleName}`],
+          ])}
+          onPress={() => toggleModal()}>
+          {(field.icon || field.iconLeft) && (
+            <FormIcon
+              name={field.icon || field.iconLeft}
+              size={field.iconSize || 20}
+              color={field.iconColor || '#000'}
+            />
+          )}
           <Text
             style={StyleSheet.flatten([
-              styles.selectPickerLabel,
-              customStyles.selectPickerLabel,
-              customStyles[`selectPickerLabel${styleName}`],
+              styles.selectPickerText,
+              customStyles.selectPickerText,
+              customStyles[`selectPickerText${styleName}`],
+              (field.icon || field.iconLeft || field.iconRight) && {
+                marginHorizontal: 5,
+              },
+              !items[valueIndex] && {
+                color: 'gray',
+                ...customStyles.selectPickerPlaceholderText,
+                ...customStyles[`selectPickerPlaceholderText${styleName}`],
+              }
             ])}>
-            {label}
+            {
+              items[valueIndex]
+              ? (items[valueIndex].label instanceof Object ? items[valueIndex].label[currentLocale] : items[valueIndex].label)
+              : placeholder
+            }
           </Text>
-        )}
+          {field.iconRight && (
+            <FormIcon
+              name={field.iconRight}
+              size={field.iconSize || 20}
+              color={field.iconColor || '#000'}
+            />
+          )}
+        </TouchableOpacity>
+      </View>
+      <Modal
+        animationType="fade"
+        transparent
+        visible={show}
+        onRequestClose={() => toggleModal()}>
+        <TouchableOpacity
+          style={StyleSheet.flatten([
+            styles.selectPickerModalOverlay,
+            customStyles.selectPickerModalOverlay,
+            customStyles[`selectPickerModalOverlay${styleName}`],
+          ])}
+          activeOpacity={1}
+          onPress={() => toggleModal()}
+        />
         <View
           style={StyleSheet.flatten([
-            styles.selectPickerTextWrapper,
-            customStyles.selectPickerTextWrapper,
-            customStyles[`selectPickerTextWrapper${styleName}`],
+            styles.selectPickerModalContainer,
+            customStyles.selectPickerModalContainer,
+            customStyles[`selectPickerModalContainer${styleName}`],
           ])}>
-          <TouchableOpacity
-            style={StyleSheet.flatten([
-              styles.selectPickerTextTouch,
-              customStyles.selectPickerTextTouch,
-              customStyles[`selectPickerTextTouch${styleName}`],
-            ])}
-            onPress={() => this.toggleModal()}>
-            {(field.icon || field.iconLeft) && (
-              <FormIcon
-                name={field.icon || field.iconLeft}
-                size={field.iconSize || 20}
-                color={field.iconColor || '#000'}
-              />
-            )}
-            <Text
-              style={StyleSheet.flatten([
-                styles.selectPickerText,
-                customStyles.selectPickerText,
-                customStyles[`selectPickerText${styleName}`],
-                (field.icon || field.iconLeft || field.iconRight) && {
-                  marginHorizontal: 5,
-                },
-                !items[valueIndex] && {
-                  color: 'gray',
-                  ...customStyles.selectPickerPlaceholderText,
-                  ...customStyles[`selectPickerPlaceholderText${styleName}`],
-                }
-              ])}>
-              {
-                items[valueIndex]
-                ? (items[valueIndex].label instanceof Object ? items[valueIndex].label[currentLocale] : items[valueIndex].label)
-                : placeholder
-              }
-            </Text>
-            {field.iconRight && (
-              <FormIcon
-                name={field.iconRight}
-                size={field.iconSize || 20}
-                color={field.iconColor || '#000'}
-              />
-            )}
-          </TouchableOpacity>
-        </View>
-        <Modal
-          animationType="fade"
-          transparent
-          visible={show}
-          onRequestClose={() => this.toggleModal()}>
-          <TouchableOpacity
-            style={StyleSheet.flatten([
-              styles.selectPickerModalOverlay,
-              customStyles.selectPickerModalOverlay,
-              customStyles[`selectPickerModalOverlay${styleName}`],
-            ])}
-            activeOpacity={1}
-            onPress={() => this.toggleModal()}
-          />
           <View
             style={StyleSheet.flatten([
-              styles.selectPickerModalContainer,
-              customStyles.selectPickerModalContainer,
-              customStyles[`selectPickerModalContainer${styleName}`],
+              styles.selectPickerModalWrapper,
+              customStyles.selectPickerModalWrapper,
+              customStyles[`selectPickerModalWrapper${styleName}`],
             ])}>
             <View
               style={StyleSheet.flatten([
-                styles.selectPickerModalWrapper,
-                customStyles.selectPickerModalWrapper,
-                customStyles[`selectPickerModalWrapper${styleName}`],
+                styles.selectPickerModalConfirmContainer,
+                customStyles.selectPickerModalConfirmContainer,
+                customStyles[`selectPickerModalConfirmContainer${styleName}`],
               ])}>
-              <View
-                style={StyleSheet.flatten([
-                  styles.selectPickerModalConfirmContainer,
-                  customStyles.selectPickerModalConfirmContainer,
-                  customStyles[`selectPickerModalConfirmContainer${styleName}`],
-                ])}>
-                <TouchableOpacity
-                  onPress={() => {
-                    this.setState({
-                      value: selectedValue,
-                      valueIndex: selectedIndex,
-                    });
-                    onChange(selectedValue);
-                    this.toggleModal('done');
-                  }}>
-                  <Text>{confirmLabel}</Text>
-                </TouchableOpacity>
-              </View>
-              <Picker
-                selectedValue={selectedValue}
-                onValueChange={(itemValue, itemIndex) => {
-                  this.setState({
-                    selectedValue: itemValue,
-                    selectedIndex: itemIndex,
-                  });
+              <TouchableOpacity
+                onPress={() => {
+                  setValue(selectedValue);
+                  setValueIndex(selectedIndex);
+                  onChange(selectedValue);
+                  toggleModal('done');
                 }}>
-                {items.map(i => {
-                  const l =
-                    i.label instanceof Object
-                      ? i.label[currentLocale]
-                      : i.label;
-                  return (
-                    <Picker.Item key={i.value} label={l} value={i.value} />
-                  );
-                })}
-              </Picker>
+                <Text>{confirmLabel}</Text>
+              </TouchableOpacity>
             </View>
+            <Picker
+              selectedValue={selectedValue}
+              onValueChange={(itemValue, itemIndex) => {
+                setSelectedValue(itemValue);
+                setSelectedIndex(itemIndex);
+              }}>
+              {items.map(i => {
+                const l =
+                  i.label instanceof Object
+                    ? i.label[currentLocale]
+                    : i.label;
+                return (
+                  <Picker.Item key={i.value} label={l} value={i.value} />
+                );
+              })}
+            </Picker>
           </View>
-        </Modal>
-        {errors[field.name] && (
-          <ErrorMessage
-            error={errors[field.name]}
-            name={styleName}
-            customStyles={customStyles}
-          />
-        )}
-      </View>
-    );
-  }
-}
+        </View>
+      </Modal>
+      {errors[field.name] && (
+        <ErrorMessage
+          error={errors[field.name]}
+          name={styleName}
+          customStyles={customStyles}
+        />
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   selectPickerContainer: {
@@ -270,7 +251,6 @@ SelectPicker.propTypes = {
   field: PropTypes.object.isRequired,
   styleName: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  reference: PropTypes.func.isRequired,
   label: PropTypes.string,
   value: PropTypes.instanceOf(Date),
   errors: PropTypes.object,
