@@ -4,45 +4,42 @@ import PropTypes from 'prop-types';
 import HTMLView from 'react-native-htmlview';
 import NestedScrollView from 'react-native-nested-scroll-view';
 
-class TextOnly extends React.PureComponent {
-  render() {
-    const {
-      label,
-      styleName,
-      customStyles,
-      description,
-    } = this.props;
-    return (
-      <View
-        style={StyleSheet.flatten([
-          styles.textOnlyContainer,
-          customStyles.textOnlyContainer,
-          customStyles[`textOnlyContainer${styleName}`],
-        ])}>
-        {label && (
-          <Text
-            style={StyleSheet.flatten([
-              styles.textOnlyLabel,
-              customStyles.textOnlyLabel,
-              customStyles[`textOnlyLabel${styleName}`],
-            ])}>
-            {label}
-          </Text>
-        )}
-        {description && (
-          <NestedScrollView
-            style={StyleSheet.flatten([
-              styles.textOnlyTextArea,
-              customStyles.textOnlyTextArea,
-              customStyles[`textOnlyTextArea${styleName}`],
-            ])}>
-            <HTMLView value={description} />
-          </NestedScrollView>
-        )}
-      </View>
-    );
-  }
-}
+const TextOnly = ({
+  label,
+  styleName,
+  customStyles,
+  description,
+}) => {
+  return (
+    <View
+      style={StyleSheet.flatten([
+        styles.textOnlyContainer,
+        customStyles.textOnlyContainer,
+        customStyles[`textOnlyContainer${styleName}`],
+      ])}>
+      {label && (
+        <Text
+          style={StyleSheet.flatten([
+            styles.textOnlyLabel,
+            customStyles.textOnlyLabel,
+            customStyles[`textOnlyLabel${styleName}`],
+          ])}>
+          {label}
+        </Text>
+      )}
+      {description && (
+        <NestedScrollView
+          style={StyleSheet.flatten([
+            styles.textOnlyTextArea,
+            customStyles.textOnlyTextArea,
+            customStyles[`textOnlyTextArea${styleName}`],
+          ])}>
+          <HTMLView value={description} />
+        </NestedScrollView>
+      )}
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   textOnlyContainer: {
